fix: dev:skip script test_ct_autoprovisionutils.py
```
TG-1060 - Skipping the test script test_ct_autoprovisionutils.py as this
script is failing and this needs to be rewritten.

Disabling this to make zuul to pass the ct.

This script is initially written to use lighttpd which is on switch, Now
lighttpd is not part of switch image is replaced by ops-restd.
Need to rewrite the script.

Tags: fix
```
Change-Id: Ie4b27e1ff9393784957ce9b79ff1c32efeaa5753
Signed-off-by: Satya Naik <satya.pra.naik@hpe.com>